requirements.txt
# Core AI/ML Libraries
langchain>=0.2.0
langchain-community>=0.2.0
openai>=1.10.0
sentence-transformers>=2.3.1
faiss-cpu==1.13.2
tiktoken>=0.5.2

# Data Processing
numpy>=2.0.0
pandas>=2.1.0
pypdf==3.17.4
python-docx==1.1.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.10.0
python-multipart>=0.0.6

# AWS Integration
boto3==1.34.34
botocore==1.34.34

# Configuration & Environment
python-dotenv==1.0.0
pyyaml>=6.0.1

# Monitoring & Logging
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3
isort==5.13.2

# Development Tools
ipython>=8.20.0
jupyter==1.0.0
